--- src/components/card/TmpCard.stories.tsx
@@ -1,36 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import Card from './Card';
-import { mockMeeting } from './mock/mock';
-
-const meta = {
-  title: 'Components/TMPCard/Base',
-  component: Card,
-  parameters: {
-    layout: 'centered',
-  },
-  argTypes: {
-    isCanceled: {
-      control: 'boolean',
-      description: '모임 취소 여부',
-    },
-  },
-} satisfies Meta<typeof Card>;
-
-export default meta;
-type Story = StoryObj<typeof Card>;
-
-export const Default: Story = {
-  args: {
-    isCanceled: false,
-  },
-  render: (args) => (
-    <Card {...args}>
-      <Card.Image {...mockMeeting.imageInfo} />
-      <Card.Box>
-        <Card.Info {...mockMeeting.meetingInfo} />
-        <Card.Status {...mockMeeting.participationStatus} />
-        <Card.Overlay onDelete={() => alert('삭제되었습니다')} />
-      </Card.Box>
-    </Card>
-  ),
-};
