--- src/components/card/types/actions.ts
@@ -1,8 +1,0 @@
-export interface SimpleActions {
-  onClick: () => void;
-  onDelete: () => void;
-}
-
-export interface FullActions {
-  onJoinClick?: () => void;
-}
