--- src/components/card/types/base.ts
@@ -1,8 +1,0 @@
-export interface BaseProps {
-  isPast?: boolean;
-  isCanceled?: boolean;
-}
-
-export interface CardContextType {
-  isCanceled: boolean;
-}
