+++ src/components/card/types/index.ts
@@ -1,5 +1,2 @@
-export * from './base';
-export * from './meeting';
-export * from './actions';
-export * from './components';
-export * from './models';
+export * from './card';
+export * from './clubCard';
