--- src/components/card/types/interface/index.ts
@@ -1,2 +1,0 @@
-export * from './card';
-export * from './clubCard';
