--- src/components/card/types/meeting.ts
@@ -1,26 +1,0 @@
-export interface MeetingInfo {
-  title: string;
-  category: string;
-  location: string;
-  datetime: string;
-}
-
-export interface ParticipantInfo {
-  src: string;
-  alt: string;
-}
-
-export interface ParticipationStatus {
-  currentParticipants: number;
-  maxParticipants: number;
-  isConfirmed: boolean;
-  confirmedVariant: 'confirmed' | 'closed';
-  participants: ParticipantInfo[];
-}
-
-export interface ImageInfo {
-  url: string;
-  alt?: string;
-  isLiked: boolean;
-  onLikeClick: () => void;
-}
